refactor: placed bacenta status cypher within arrivals folder
